Change high-contrast indeterminate progress overlay color
The native theme has no border on an indeterminate progress bar and for
unknown reasons fills the background with the highlight color which
normally would be used for the moving overlay. The overlay instead has a
hardcoded color of #FF00D228. In a bad case this results in mismatched
colors, in the worst case low-contrast or even an invisible overlay. We
instead mimick the behavior of the non-highcontrast aero themes, fill the
overlay with the highlight color and keep the border.