UxThemeEx: Load internal uxtheme symbols from debug symbols
ThemePreviewer requires access to certain uxtheme internals to properly
display and refresh styles. These symbol addresses have been hardcoded
until now and required manual changes whenever uxtheme.dll was updated.

From now on we attempt to get these symbols using the provided debugging
symbols. If this fails, we fallback to a few hardcoded addresses for
certain versions. If even this fails, do not use these internals. This
will result in display errors but not crash the program.

What's missing from this is notifying the user if the internals cannot be
used.