Fix TabItem height when there is only a single item
Previously the TabItem height would be 1px smaller when alone compared to
to a TabControl with multiple items. Due to the weird behavior with
negative margins adjusting the padding on the outer border results in
either a 0px or 2px adjustment. Moving the padding to a second nested
border properly applies a 1px adjustment.