All themes: Paint top-level menu items pressed when keyboard focused
When a top-level menu item (TopLevelHeaderTemplateKey) is focused with the
keyboard (e.g. after pressing Alt or after dismissing an opened top-level
menu popup) it must be painted as pressed to indicate the currently
selected when navigating with the left or right arrows.

Using IsKeyboardFocused instead of IsPressed for the style triggers handles
both keyboard and mouse focused states.

While at it, remove the superfluous IsSubmenuOpen trigger for top-level
items since they do not have a submenu.